+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ### v0.1.27
-- Callback : Allow modifying `result`
-- Instance : Allow calling object functions with `:`
+- Callback : Allow modifying `result` via return value
 - SurvivorLog : Fix "Max Health" not displaying correctly when calling `new_from_survivor`
+- Instance : Allow calling object functions with `:`
 
 ### v0.1.26
 - Survivor : Fix error thrown on death for custom survivors